✨ feat: Separate files that write use client - Separate interactive elements into client components - Separate other elements into server components
